+++ LearnositySDKUnitTests/InitTests.cs
@@ -48,11 +48,11 @@ namespace LearnositySDKUnitTests
         private Dictionary<string, string> getInitTestCases() {
             return new Dictionary<string, string>
             {
-                { "assess", "03f4869659eeaca81077785135d5157874f4800e57752bf507891bf39c4d4a90" },
-                { "author", "108b985a4db36ef03905572943a514fc02ed7cc6b700926183df7babc2cd1c96" },
-                { "items", "dd6bf2a5fd28c9935acef5e2918a1069269154414c19ab346d476c363a7a964c" },
-                { "questions", "03f4869659eeaca81077785135d5157874f4800e57752bf507891bf39c4d4a90" },
-                { "reports", "91085beccf57bf0df77c89df94d1055e631b36bc11941e61460b445b4ed774bc" },
+                { "assess", "$02$8de51b7601f606a7f32665541026580d09616028dde9a929ce81cf2e88f56eb8" },
+                { "author", "$02$ca2769c4be77037cf22e0f7a2291fe48c470ac6db2f45520a259907370eff861" },
+                { "items", "$02$78ef0334e708829bd3c92decc040f91c8433b07ba32cc9198705a18c36c2ea54" },
+                { "questions", "$02$8de51b7601f606a7f32665541026580d09616028dde9a929ce81cf2e88f56eb8" },
+                { "reports", "$02$8e0069e7aa8058b47509f35be236c53fa1a878c64b12589fd42f48b568f6ac84" },
             };
         }
 
@@ -60,7 +60,7 @@ namespace LearnositySDKUnitTests
         public void testGenerateJsonDataApiGetRequest()
         {
             Init init = TestRequest.getTestRequestFor("data", "get").getJsonInit();
-            string expectedSignature = "e1eae0b86148df69173cb3b824275ea73c9c93967f7d17d6957fcdd299c8a4fe";
+            string expectedSignature = "%2402%24e19c8a62fba81ef6baf2731e2ab0512feaf573ca5ca5929c2ee9a77303d2e197";
 
 
             Assert.AreEqual(
@@ -73,7 +73,7 @@ namespace LearnositySDKUnitTests
         public void testGenerateJsonDataApiPostRequest()
         {
             Init init = TestRequest.getTestRequestFor("data", "post").getJsonInit();
-            string expectedSignature = "18e5416041a13f95681f747222ca7bdaaebde057f4f222083881cd0ad6282c38";
+            string expectedSignature = "%2402%249d1971fb9ac51482f7e73dcf87fc029d4a3dfffa05314f71af9d89fb3c2bcf16";
 
             Assert.AreEqual(
                 "security=%7b%22consumer_key%22%3a%22yis0TYCu7U9V4o7M%22%2c%22domain%22%3a%22localhost%22%2c%22timestamp%22%3a%2220140626-0528%22%2c%22signature%22%3a%22" + expectedSignature + "%22%7d&request=%7b%22limit%22%3a100%7d&action=post",